requirements.txt
pulsar-client==2.10.0
protobuf==4.22.0
